+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## next
+
+- Fixed wrong "Converting circular structure to JSON" error when empty object or array is reused in object to stringify
+
 ## 1.8.1 (23-04-2020)
 
 - Fixed `--prebuild` option for server to not prebuild when option is not set